Refactor : header 컴포넌트의 인터페이스 파일명을 header.ts 로 수정
